scripts/video_automation/requirements.txt
elevenlabs
moviepy
python-dotenv
google-genai
requests
